refactor(logging): consolidate 4 overlapping layers into RunLog+Logfire
PR #36 added Logfire/OTel telemetry that subsumes the legacy fallbacks:
- Logfire auto-attaches the stdlib logging tree, making PythonLoggingSink
  a double-emit.
- OTel's active span carries run_id/issue_number/persona/phase
  implicitly, so the _WATCHER_LOGGER global + use_watcher_fallback branch
  are dead.
- logging_config.py was 16 lines chaining two inits.

Cuts:
- Delete PythonLoggingSink (Logfire bridges stdlib)
- Delete _WATCHER_LOGGER singleton + use_watcher_fallback in
  emit_fleet_event; route ambient events to logfire.{level}() under the
  active span instead.
- Delete logging_config.py; fold configure_fleet_logging into
  telemetry.py and update callers.

Keeps: FleetEvent (canonical record), JsonlFileSink (per-run forensic),
MemoryRingSink (status queries), LogfireSink, RunLog, FleetLogger,
context.py helpers.
