--- agent_fleet/logging_config.py
@@ -1,16 +1,0 @@
-"""Shared logging setup for fleet CLIs and watchers."""
-
-from __future__ import annotations
-
-import logging
-
-from agent_fleet.telemetry import configure_telemetry
-
-
-def configure_fleet_logging(level: int = logging.INFO) -> None:
-    logging.basicConfig(
-        level=level,
-        format="%(asctime)s %(levelname)s %(name)s: %(message)s",
-        force=True,
-    )
-    configure_telemetry()
