+++ agent_fleet/observability/__init__.py
@@ -2,18 +2,9 @@
 
 from agent_fleet.observability.context import bind_run, get_run_context, get_run_log
 from agent_fleet.observability.events import FleetEvent, RunContext
-from agent_fleet.observability.fleet_logger import (
-    FleetLogger,
-    emit_fleet_event,
-    get_watcher_logger,
-)
+from agent_fleet.observability.fleet_logger import FleetLogger, emit_fleet_event
 from agent_fleet.observability.log import RunLog
-from agent_fleet.observability.sinks import (
-    JsonlFileSink,
-    LogSink,
-    MemoryRingSink,
-    PythonLoggingSink,
-)
+from agent_fleet.observability.sinks import JsonlFileSink, LogSink, MemoryRingSink
 
 __all__ = [
     "FleetEvent",
@@ -21,12 +12,10 @@ __all__ = [
     "JsonlFileSink",
     "LogSink",
     "MemoryRingSink",
-    "PythonLoggingSink",
     "RunContext",
     "RunLog",
     "bind_run",
     "emit_fleet_event",
     "get_run_context",
     "get_run_log",
-    "get_watcher_logger",
 ]
